실행 최소 가능하게 변경

+++ src/main/java/com/spharosacademy/project/SSGBack/order/dto/request/OrdersOptioninputDto.java
@@ -15,6 +15,6 @@ public class OrdersOptioninputDto {
 
     private Long optionListId;
     private String orderMsg;
-    private long qty;
+    private int qty;
 
 }
